modelo y controlador de especialidad

+++ database/migrations/2024_01_29_215322_create_doctor_speciality_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('doctor_specialties', function (Blueprint $table) {
             $table->id();
-            $table->unsignedBigInteger('doctor_id');
+            $table->unsignedInteger('doctor_id');
             $table->unsignedBigInteger('speciality_id');
 
             $table->foreign('doctor_id')->references('id')->on('doctors');
